[feat] fix ProgressBar file structure
